Task/upgrade model (#367)
* Updated tools project to newest version of model package
* Changed type map to use underlying type of enum instead of long

+++ tools/HatTrick.DbEx.Tools/Model/ColumnModel.cs
@@ -30,7 +30,7 @@ namespace HatTrick.DbEx.Tools.Model
         public bool IsNullable { get; }
         public bool IsIdentity { get; }
         public bool IsComputed { get; }
-        public short? Size { get; }
+        public long? Size { get; }
         public byte? Precision { get; }
         public byte? Scale { get; }
         public IDictionary<string, string> Properties { get; }
